Move translation endpoints into a separate file
